refactor: address review — neutral detection module, cheaper skip path, --runs validation
- move rules/seo/detection.ts to rules/detection.ts: the shared PASS/
  PENALIZED constants are consumed by the generic file engine and the
  correctness/perf rules, so they no longer live under an SEO directory
- resolve a fileRule's recommendation only after the applies guard, so
  function-valued recommendations don't run for skipped files
- bench.mjs: reject non-positive-integer --runs instead of silently
  running zero times
