+++ packages/core/src/rules/detection.ts
