+++ packages/core/src/rules/seo/jsonld-engine.ts
@@ -2,7 +2,7 @@
 import type { Fix, Result } from '../../types.js';
 import { docsUrlFor, type Rule, type RuleContext } from '../../rule.js';
 import type { HeadTag } from '../../head.js';
-import { PENALIZED, PASS } from './detection.js';
+import { PENALIZED, PASS } from '../detection.js';
 
 export type JsonLdNode = Record<string, unknown>;
